requirements.txt
aiosqlite==0.21.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.3.0
bleach==6.1.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.114.2
fastapi-jwt-auth==0.5.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==1.7.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.1.1
python-jose==3.3.0
python-slugify==8.0.4
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
text-unidecode==1.3
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1